+++ src/operator/subgraph/partition_graph.cc
@@ -30,6 +30,7 @@
 #include <stack>
 
 #include "./subgraph_op.h"
+#include "./common.h"
 
 namespace nnvm {
 NodePtr CreateVariableNode(const std::string& name);
@@ -45,26 +46,45 @@ using nnvm::NodePtr;
 using nnvm::NodeEntry;
 using nnvm::Graph;
 
-// use the function in quantize_graph.cc
-NodePtr CreateNode(std::string, std::string);
-NodePtr CloneVariableNode(const nnvm::Node& src) {
-  CHECK(src.is_variable());
-  CHECK_EQ(src.inputs.size(), 0U);
-  CHECK_EQ(src.control_deps.size(), 0U);
-  NodePtr node = nnvm::Node::Create();
-  *node = src;
-  return node;
-}
+// TODO(junwu): Change this to 0
+#define SUBGRAPH_DEBUG 1
 
 namespace sg {  // sg stands for subgraph
 
+#if SUBGRAPH_DEBUG
+void PrintSubgraph(const std::vector<SimpleNode*>& simple_nodes) {
+  std::string op_names = "";
+  for (size_t i = 0; i < simple_nodes.size(); ++i) {
+    op_names += simple_nodes[i]->node->attrs.name + ' ';
+  }
+  LOG(INFO) << "Subgraph node names: " << op_names;
+}
+
+void PrintNodeEntry(const nnvm::NodeEntry& entry) {
+  std::string ret = "NodeEntry: node_name=" + entry.node->attrs.name
+    + ", index=" + std::to_string(entry.index) + ", version=" + std::to_string(entry.version);
+  LOG(INFO) << ret;
+}
+
+void PrintNodeEntries(const std::vector<nnvm::NodeEntry*>& entries) {
+  for (size_t i = 0; i < entries.size(); ++i) {
+    PrintNodeEntry(*entries[i]);
+  }
+}
+#endif
+
+/*!
+ * \brief Given a MXNet computational graph, create an undirected graph from it.
+ * \param g the MXNet computational graph
+ * \param simple_nodes the nodes of undirected graph in top sorted order
+ */
 void CreateSimpleGraph(const Graph& g,
                        std::vector<SimpleNodePtr>* simple_nodes) {
   const auto& indexed_graph = g.indexed_graph();
   simple_nodes->reserve(indexed_graph.num_nodes());
-  for (size_t nid = 0; nid < indexed_graph.num_nodes(); ++nid) {
+  DFSVisit(g.outputs, [&](const NodePtr& node) {
     SimpleNodePtr sn = SimpleNode::Create();
-    sn->node = const_cast<nnvm::Node*>(indexed_graph[nid].source);
+    sn->node = node.get();
     for (size_t i = 0; i < sn->node->inputs.size(); ++i) {
       const auto& e = sn->node->inputs[i];
       const auto input_nid = indexed_graph.node_id(e.node.get());
@@ -78,29 +98,13 @@ void CreateSimpleGraph(const Graph& g,
       }
     }
     simple_nodes->emplace_back(std::move(sn));
-  }
-#if 0
-  DFSVisit(g.outputs, [&](const NodePtr& node) {
-    LOG(INFO) << node->attrs.name;
-    auto it = node_map2->find(node.get());
-    if (it == node_map2.end()) {
-    }
-
-    for (const auto& e : node->inputs) {
-      LOG(INFO) << "NodeEntry: " << e.node->attrs.name << ", " << e.index << ", " << e.version;
-    }
   });
-#endif
 }
 
-void PrintSubgraph(const std::vector<SimpleNode*>& simple_nodes) {
-  std::string op_names = "";
-  for (size_t i = 0; i < simple_nodes.size(); ++i) {
-    op_names += simple_nodes[i]->node->attrs.name + ' ';
-  }
-  LOG(INFO) << "Subgraph node names: " << op_names;
-}
-
+/*!
+ * \brief Reset labels of the subgraph nodes to the original state
+ * and clear the vector of subgraph nodes.
+ */
 void ResetSubgraphNodes(std::vector<SimpleNode*>* subgraph_nodes) {
   for (auto sn : *subgraph_nodes) {
     sn->label = -1;
@@ -108,13 +112,20 @@ void ResetSubgraphNodes(std::vector<SimpleNode*>* subgraph_nodes) {
   subgraph_nodes->clear();
 }
 
-/*
- * This function traverses the nodes in a computation graph from a starting
- * node following the input links and output links, and marks all nodes that
+/*!
+ * \brief This function traverses the nodes in a computation graph from a starting
+ * node following the input edges and output edges, and marks all nodes that
  * can be accessed from the starting node. Before the function returns,
  * it will conduct checking whether there is a loop between the potential subgraph
  * and the outside nodes. If so, add the node that should break the loop
  * in excluded_nodes and return false. Otherwise, return true.
+ * \param g the whole graph
+ * \select_func determines whether the visited node should be choosen or not
+ * \label the label of the current subgraph
+ * \snid node id of the seed simple node
+ * \simple_nodes all simple nodes in the top sorted order
+ * \subgraph_nodes all the nodes belonging to the same subgraph of seed node
+ * \excluded_nodes set of nodes that should be excluded from the current subgraph
  */
 bool LabelSubgraph(const Graph& g,
                    SubgraphSelectorPtr select_func,
@@ -129,12 +140,14 @@ bool LabelSubgraph(const Graph& g,
     node_queue.push(simple_nodes[snid].get());
   }
   // key: nodes that serve as input/output nodes to the subgraph
-  // value: pair of vectors of nodes in the subgraph. The first vector contains the output nodes of the key in the subgraph,
-  // and the second vector contains the input ndoes of the key in the subgraph.
-  // If both vectors are non-empty, it means there is a loop between the subgraph and the key node.
+  // value: pair of vectors of nodes in the subgraph. The first vector contains the
+  // output nodes of the key in the subgraph, and the second vector contains the
+  // input ndoes of the key in the subgraph. If both vectors are non-empty,
+  // it means there is a loop between the subgraph and the key node.
   // When breaking the loop, we want to start removing the node with the largest node id.
   std::unordered_map<const nnvm::Node*,
-    std::pair<std::vector<const nnvm::Node*>, std::vector<const nnvm::Node*>>> non_subgraph_node_map;
+    std::pair<std::vector<const nnvm::Node*>,
+              std::vector<const nnvm::Node*>>> non_subgraph_node_map;
   while (!node_queue.empty()) {
     SimpleNode* cur_node = node_queue.front();
     node_queue.pop();
@@ -205,10 +218,19 @@ bool LabelSubgraph(const Graph& g,
   return true;
 }
 
+/*!
+ * \brief Finds all the nodes belonging to the same subgraph given a seed node.
+ * \param g the whole graph
+ * \select_func determines whether the visited node should be choosen or not
+ * \label the label of the current subgraph
+ * \snid node id of the seed simple node
+ * \simple_nodes all simple nodes in the top sorted order
+ * \subgraph_nodes all the nodes belonging to the same subgraph of seed node
+ */
 void FindSubgraph(const Graph& g,
                   SubgraphSelectorPtr select_func,
                   const int label,
-                  const size_t snid,  // simple node id, this is a seed
+                  const size_t snid,
                   const std::vector<SimpleNodePtr>& simple_nodes,
                   std::vector<SimpleNode*>* subgraph_nodes) {
   std::unordered_set<const nnvm::Node*> excluded_nodes;
@@ -216,7 +238,8 @@ void FindSubgraph(const Graph& g,
   size_t count = 0;
   bool success = false;
   while (!success && count < max_num_retry) {
-    success = LabelSubgraph(g, select_func, label, snid, simple_nodes, subgraph_nodes, &excluded_nodes);
+    success = LabelSubgraph(g, select_func, label, snid, simple_nodes,
+                            subgraph_nodes, &excluded_nodes);
     if (!success) {
       CHECK(!excluded_nodes.empty());
       std::string excluded_node_names;
@@ -239,11 +262,9 @@ void FindSubgraph(const Graph& g,
   }
 }
 
-/*
- * This function finds subgraphs with all nodes that meet certain criteria.
+/*!
+ * \brief Finds subgraphs with all nodes that meet certain criteria.
  * All nodes in a subgraph are marked with the same label.
- * All nodes in a subgraph have to be connected with each other. If a node
- * doesn't meet the given criteria, it will be marked with a separate label.
  */
 void FindSubgraphs(const Graph& g,
                    const SubgraphProperty &subg_prop,
@@ -264,7 +285,7 @@ void FindSubgraphs(const Graph& g,
 }
 
 /*!
- * \brief Sort entries according to their topological order.
+ * \brief Sorts entries according to their topological order.
  * Note that entry ids cannot be used to sort entries.
  * \param entry_top_order_map mapping from entry pointer to its topological position in the graph
  * \param entries Node entries to be sorted
@@ -281,11 +302,15 @@ void SortEntries(const std::unordered_map<const nnvm::NodeEntry*, size_t>& entry
   std::sort(entries->begin(), entries->end(), entry_cmp);
 }
 
-/*
- * \brief find the input entries of a subgraph
- * \param input_entry_map mapping from node entry pointer to the
- * pair of its dest node and index in the dest node's inputs
+/*!
+ * \brief Given a subgraph, find the output entries of a subgraph.
+ * \param g pointer to the whole graph
+ * \param simple_nods vector of simple nodes in top sorted order
+ * \param subgraph_nodes vector of pointers of simples of a subgraph.
+ * \param entry_top_order_map mapping entry pointer to its top sorted position
+ * \param input_entries input entries of the subgraph
  */
+
 void FindInputEntries(const Graph& g,
                       const std::vector<SimpleNodePtr>& simple_nodes,
                       const std::vector<SimpleNode*>& subgraph_nodes,
@@ -319,7 +344,14 @@ void FindInputEntries(const Graph& g,
   SortEntries(entry_top_order_map, input_entries);
 }
 
-// find the output entries of a subgraph
+/*!
+ * \brief Given a subgraph, find the output entries of a subgraph.
+ * \param g pointer to the whole graph
+ * \param simple_nods vector of simple nodes in top sorted order
+ * \param subgraph_nodes vector of pointers of simples of a subgraph.
+ * \param entry_top_order_map mapping entry pointer to its top sorted position
+ * \param output_entries output entries of the subgraph
+ */
 void FindOutputEntries(Graph* g,
                        const std::vector<SimpleNodePtr>& simple_nodes,
                        const std::vector<SimpleNode*>& subgraph_nodes,
@@ -374,19 +406,7 @@ void FindOutputEntries(Graph* g,
   SortEntries(entry_top_order_map, output_entries);
 }
 
-void PrintNodeEntry(const nnvm::NodeEntry& entry) {
-  std::string ret = "NodeEntry: node_name=" + entry.node->attrs.name
-    + ", index=" + std::to_string(entry.index) + ", version=" + std::to_string(entry.version);
-  LOG(INFO) << ret;
-}
-
-void PrintNodeEntries(const std::vector<nnvm::NodeEntry*>& entries) {
-  for (size_t i = 0; i < entries.size(); ++i) {
-    PrintNodeEntry(*entries[i]);
-  }
-}
-
-/*
+/*!
  * \brief Given a computation graph and a set of input node entries, this function cuts
  * the node entries and creates new variable nodes as the input nodes of the
  * subgraph. It returns the nodes that connect to the subgraph directly and
@@ -413,23 +433,28 @@ void CutGraphInputs(const std::vector<nnvm::NodeEntry*> &input_entries,
   }
 }
 
-// Replace a set of nodes belonging to the same subgraph with a subgrpah node
-// and keep the subgraph in the subgraph node. The input entries and output entries
-// of the subgraph node are kept in the same order as the subgraph's.
+/*!
+ * \brief Replace a set of nodes belonging to the same subgraph with a subgrpah node
+ * and keep the subgraph in the subgraph node. The input entries and output entries
+ * of the subgraph node are kept in the same order as the subgraph's.
+ */
 void CreateSubgraphNode(Graph* g,
                         const std::vector<SimpleNodePtr>& simple_nodes,
                         const std::vector<SimpleNode*>& subgraph_nodes,
                         const size_t subgraph_id,
                         std::unordered_map<const nnvm::NodeEntry*, size_t>* entry_top_order_map) {
+#if SUBGRAPH_DEBUG
   LOG(INFO) << "Searching for input entries...";
+#endif
   std::vector<nnvm::NodeEntry*> input_entries;
   FindInputEntries(*g, simple_nodes, subgraph_nodes, *entry_top_order_map, &input_entries);
   std::vector<nnvm::NodeEntry> orig_input_entries;
   // TODO(junwu): Confirm what value to pass to skip_var
   CutGraphInputs(input_entries, &orig_input_entries, false);
+#if SUBGRAPH_DEBUG
   PrintNodeEntries(input_entries);
-
   LOG(INFO) << "Searching for output entries...";
+#endif
   std::vector<nnvm::NodeEntry*> output_entries;
   FindOutputEntries(g, simple_nodes, subgraph_nodes, *entry_top_order_map, &output_entries);
 
@@ -466,7 +491,9 @@ void CreateSubgraphNode(Graph* g,
       sn->outputs[n.get()].push_back(i);
     }
   }
+#if SUBGRAPH_DEBUG
   PrintNodeEntries(output_entries);
+#endif
 }
 
 }  // namespace sg
@@ -476,10 +503,11 @@ void CreateSubgraphNode(Graph* g,
  * This is going to be used to sort input/output entries of subgraphs to keep
  * the topological order unchanged.
  */
-void TopSortEntries(const Graph& g, std::unordered_map<const nnvm::NodeEntry*, size_t>* entry_top_order_map) {
+void TopSortEntries(const Graph& g,
+                    std::unordered_map<const nnvm::NodeEntry*, size_t>* entry_top_order_map) {
   CHECK(entry_top_order_map != nullptr);
   std::unordered_set<const nnvm::Node*> visited;
-  // meaning of tuple: (graph node, index of node's inputs, node entry as the output of the graph node)
+  // tuple: (graph node, index of node's inputs, node entry as the output of the graph node)
   std::stack<std::tuple<nnvm::Node*, size_t, const nnvm::NodeEntry*>> s;
   auto in_degree = [] (const nnvm::Node* node)->size_t {
     if (!node) {
@@ -521,76 +549,36 @@ void TopSortEntries(const Graph& g, std::unordered_map<const nnvm::NodeEntry*, s
 }
 
 Graph PartitionGraph(Graph&& g) {
-#if 0
-  DFSVisit(g.outputs, [&](const NodePtr& node) {
-    LOG(INFO) << node->attrs.name;
-    for (const auto& e : node->inputs) {
-      LOG(INFO) << "NodeEntry: " << e.node->attrs.name << ", " << e.index << ", " << e.version;
-    }
-  });
-#endif
   if (!g.HasAttr("subgraph_property")) {  // treat the whole graph as a subgraph
-    Symbol whole_graph_sym;
-    whole_graph_sym.outputs = g.outputs;
-    // DO NOT define node name for subgraph op because it would serve
-    // as the prefix of the output names. We want to use the original
-    // output names of the subgraph.
-    NodePtr subgraph_node_ptr = CreateNode("_subgraph_op", "");
-    subgraph_node_ptr->attrs.parsed = whole_graph_sym;
-    const auto& idx = g.indexed_graph();
-    const auto& input_node_ids = idx.input_nodes();
-    for (size_t i = 0; i < input_node_ids.size(); ++i) {
-      const auto& input_node = idx[input_node_ids[i]];
-      // also need to clone the attrs of the source variable node
-      NodePtr new_input_node = CloneVariableNode(*input_node.source);
-      CHECK_EQ(new_input_node->inputs.size(), 0U);
-      CHECK_EQ(subgraph_node_ptr->inputs.size(), i);
-      nnvm::NodeEntry new_node_entry{new_input_node, 0, 0};
-      subgraph_node_ptr->inputs.emplace_back(new_node_entry);
-    }
-    Graph ret;
-    static const auto& flist_outputs = nnvm::Op::GetAttr<nnvm::FListOutputNames>("FListOutputNames");
-    auto list_output_names_func = flist_outputs.get(subgraph_node_ptr->op(), nullptr);
-    CHECK(list_output_names_func != nullptr);
-    const size_t num_outputs = list_output_names_func(subgraph_node_ptr->attrs).size();
-    for (uint32_t i = 0; i < num_outputs; ++i) {
-      ret.outputs.emplace_back(nnvm::NodeEntry{subgraph_node_ptr, i, 0});
-    }
-    return ret;
-  } else {
-    using namespace sg;
-    const SubgraphPropertyPtr& subg_prop = g.GetAttr<SubgraphPropertyPtr>("subgraph_property");
-    // top sort NodeEntry of all the nodes' inputs
-    std::unordered_map<const nnvm::NodeEntry*, size_t> entry_top_order_map;
-    TopSortEntries(g, &entry_top_order_map);
-
-    // Create undirected graph for ease of finding subgraphs
-    std::vector<SimpleNodePtr> simple_nodes;
-    CreateSimpleGraph(g, &simple_nodes);
-#if 0
-    const auto& indexed_graph = g.indexed_graph();
-    for (size_t i = 0; i < indexed_graph.num_nodes(); ++i) {
-      const nnvm::Node* node = indexed_graph[i].source;
-      LOG(INFO) << node->attrs.name;
-    }
-#endif
-    std::vector<std::vector<SimpleNode*>> subgraph_nodes;
-    FindSubgraphs(g, *subg_prop, simple_nodes, &subgraph_nodes);
-    for (size_t i = 0; i < subgraph_nodes.size(); ++i) {
-#if 0
-      // TODO(junwu): for debugging purpose, delete the following two lines
-      std::set<SimpleNode*> simple_node_set(subgraph_nodes[i].begin(), subgraph_nodes[i].end());
-      CHECK_EQ(simple_node_set.size(), subgraph_nodes[i].size());
-#endif
-      PrintSubgraph(subgraph_nodes[i]);
-      CreateSubgraphNode(&g, simple_nodes, subgraph_nodes[i], i, &entry_top_order_map);
-    }
+    LOG(INFO) << "The graph has no attribute of subgraph_property attached. "
+                 "The original graph is returned.";
     return g;
   }
+  using namespace sg;
+  const SubgraphPropertyPtr& subg_prop = g.GetAttr<SubgraphPropertyPtr>("subgraph_property");
+  // top sort NodeEntry of all the nodes' inputs
+  std::unordered_map<const nnvm::NodeEntry*, size_t> entry_top_order_map;
+  TopSortEntries(g, &entry_top_order_map);
+
+  // Create undirected graph for ease of finding subgraphs
+  std::vector<SimpleNodePtr> simple_nodes;
+  CreateSimpleGraph(g, &simple_nodes);
+  std::vector<std::vector<SimpleNode*>> subgraph_nodes;
+  FindSubgraphs(g, *subg_prop, simple_nodes, &subgraph_nodes);
+  for (size_t i = 0; i < subgraph_nodes.size(); ++i) {
+#if SUBGRAPH_DEBUG
+    std::set<SimpleNode*> simple_node_set(subgraph_nodes[i].begin(), subgraph_nodes[i].end());
+    CHECK_EQ(simple_node_set.size(), subgraph_nodes[i].size());
+    PrintSubgraph(subgraph_nodes[i]);
+#endif
+    CreateSubgraphNode(&g, simple_nodes, subgraph_nodes[i], i, &entry_top_order_map);
+  }
+  return g;
 }
 
 NNVM_REGISTER_PASS(PartitionGraph)
-.describe("")
+.describe("Partition a graph according to the user defined rules "
+          "in a derived class of SubgraphProperty")
 .set_body(PartitionGraph)
 .set_change_graph(true);
